+++ Changelog.md
@@ -47,6 +47,7 @@ v2.37.0 (not published yet)
 
 - UI: Enhance Auto Crop for Dolby Vision sources
 - UI: Improve handling with Dolby Vision videos and manual cropping
+- UI: Re-read modified Dolby Vision metadata automatically on window focus
 - UI: Improve Dolby Vision metadata handling
 - UI: Enhance "Frame Mismatch" error message
 - UI: Let "Frame Mismatch" ignore cutted "Copy/Mux" runs ([#1234](/../../issues/1234))